Kitchen doctor.
Refactor seperate resource/provider into 12.5 custom resource style
