+++ .kitchen.yml
@@ -10,8 +10,12 @@ driver:
   elevated_password: vagrant
 
 provisioner:
-  product_name: chefdk
+  product_name: chef
+  deprecations_as_errors: true
+  enforce_idempotency: true
+  multiple_converge: 2
   product_version: 13.8.5
+  install_strategy: once
 
 platforms:
   - name: windows-2012R2
